rename back main class, to make east diff
